deduplicate test with broken shadow disk, fix test
