--- cloud/disk_manager/internal/pkg/facade/image_service_nemesis_nrd_test/ya.make
@@ -1,12 +1,0 @@
-GO_TEST_FOR(cloud/disk_manager/internal/pkg/facade)
-
-SET_APPEND(RECIPE_ARGS --nemesis)
-SET_APPEND(RECIPE_ARGS --disk-agent-count 3)
-SET_APPEND(RECIPE_ARGS --retry-broken-disk-registry-based-disk-checkpoint)
-INCLUDE(${ARCADIA_ROOT}/cloud/disk_manager/internal/pkg/facade/testcommon/common.inc)
-
-GO_XTEST_SRCS(
-    ../image_service_nrd_test/image_service_nrd_test.go
-)
-
-END()
