+++ cloud/disk_manager/internal/pkg/facade/snapshot_image_common_nemesis_test/ya.make
@@ -1,12 +1,12 @@
 GO_TEST_FOR(cloud/disk_manager/internal/pkg/facade)
 
-SET_APPEND(RECIPE_ARGS --creation-and-deletion-allowed-only-for-disks-with-id-prefix "Test")
+SET_APPEND(RECIPE_ARGS --nemesis)
 SET_APPEND(RECIPE_ARGS --disk-agent-count 3)
 SET_APPEND(RECIPE_ARGS --retry-broken-disk-registry-based-disk-checkpoint)
 INCLUDE(${ARCADIA_ROOT}/cloud/disk_manager/internal/pkg/facade/testcommon/common.inc)
 
 GO_XTEST_SRCS(
-    snapshot_service_nrd_test.go
+    ../snapshot_image_common_test/snapshot_image_common_test.go
 )
 
 END()
