+++ cloud/disk_manager/internal/pkg/facade/ya.make
@@ -25,6 +25,8 @@ RECURSE_FOR_TESTS(
     placement_group_service_test
     private_service_nemesis_test
     private_service_test
+    snapshot_image_common_test
+    snapshot_image_common_nemesis_test
     snapshot_service_nemesis_test
     snapshot_service_test
     testcommon
